pkgs/tvl: expose tvl's depot in vuizvui
depot is a nix-based monorepo which contains some great nix utilities
like yants (a nix type system), runTestsuite, mergePatch and so on, a
few interesting pure nix builders like buildLisp and buildGo and a few
packages maintained by @Profpatsch and myself.

This change exposes tvl completely as pkgs.tvl, but prevents hydra from
building it using dontRecurseIntoAttrs as depot pins its own version of
nixpkgs which is not easily overrideable, contains some expensive to
build system configurations we are not interested in and even some
notoriously indeterministic packages.

Additionally it is possible to override pkgs.tvl to use a different or
local version of depot:

    pkgs.tvl.override { tvlSrc = /home/lukas/src/depot; }

Since depot is pinned to a specific revision (and there are no force
pushes) as well as depot's nixpkgs there should be no breakage caused by
depot on its own.

Downsides of the current approach:

* Hydra only builds tvl packages if they are otherwise used (in vuizvui
  pkgs or machines). Probably an upside rather than a downside.
* builtins.fetchGit

Alternatives:

* Vendoring in depot. Probably not desireable as it would explode
  vuizvuis size.
* Only exposing a subset of tvl, but also exposing it to hydra.
* Have a <tvl> channel?
* Vendor in stuff we need and keep it updated manually.
* f̵l̵a̵k̵e̵s̵

Future work:

* Get tvl to generate tarballs for git revisions so we can use
  builtins.fetchTarball and pin a hash instead of relying on fetchGit's
  caching behavior (I think fetchGit is fine and caches infinitely if a
  revision is set, but I don't know for sure as it is not documented)
* Make globally overriding tvlSrc more convenient either via an entry
  in the vuizvui config (like for pkgs.games credentials) or via some
  kind of vuizvui overlay support.
* Generally improve the story of including nix-based monorepos without
  flakes by expanding on stuff like
  <https://code.tvl.fyi/about/docs/designs/SPARSE_CHECKOUTS.md>
  and making it actually usable (with nix).